refactor: comprehensive code quality improvements
- Remove unused visitor parameter from NodeVisitorDelegate
- Implement dynamic version loading system with Version.kt
- Split monolithic DSL files into focused modules:
  * HoverDsl.kt (516 lines) → hover/ package (4 files)
  * CompletionDsl.kt (388 lines) → completion/ package (3 files)
- Add method-level @Suppress annotations for generic exception handling
- Fix circular dependencies and update all imports
- Maintain backward compatibility while improving organization

All tests passing, lint clean, build successful.
