+++ src/main/kotlin/com/github/albertocavalcante/groovylsp/dsl/hover/HoverNodeConverters.kt
@@ -1,4 +1,4 @@
-package com.github.albertocavalcante.groovylsp.dsl
+package com.github.albertocavalcante.groovylsp.dsl.hover
 
 import com.github.albertocavalcante.groovylsp.errors.LspResult
 import com.github.albertocavalcante.groovylsp.errors.toLspResult
@@ -19,99 +19,6 @@ import org.codehaus.groovy.ast.expr.GStringExpression
 import org.codehaus.groovy.ast.expr.MethodCallExpression
 import org.codehaus.groovy.ast.expr.VariableExpression
 import org.eclipse.lsp4j.Hover
-import org.eclipse.lsp4j.MarkupContent
-import org.eclipse.lsp4j.MarkupKind
-
-/**
- * Sealed class hierarchy for different types of hover content
- */
-sealed class HoverContent {
-    data class Text(val value: String) : HoverContent()
-    data class Code(val value: String, val language: String = "groovy") : HoverContent()
-    data class Markdown(val value: String) : HoverContent()
-    data class Section(val title: String, val content: kotlin.collections.List<HoverContent>) : HoverContent()
-    data class List(val items: kotlin.collections.List<String>) : HoverContent()
-    data class KeyValue(val pairs: kotlin.collections.List<Pair<String, String>>) : HoverContent()
-}
-
-/**
- * Renders HoverContent to markdown string
- */
-fun HoverContent.render(): String = when (this) {
-    is HoverContent.Text -> value
-    is HoverContent.Code -> "```$language\n$value\n```"
-    is HoverContent.Markdown -> value
-    is HoverContent.Section -> {
-        val contentStr = content.joinToString("\n\n") { item: HoverContent -> item.render() }
-        "### $title\n\n$contentStr"
-    }
-    is HoverContent.List -> items.joinToString("\n") { "- $it" }
-    is HoverContent.KeyValue -> pairs.joinToString("\n") { "**${it.first}**: ${it.second}" }
-}
-
-/**
- * Builder for creating hover content using DSL
- */
-class HoverBuilder {
-    private val content = mutableListOf<HoverContent>()
-
-    fun text(value: String) {
-        content += HoverContent.Text(value)
-    }
-
-    fun code(language: String = "groovy", value: String) {
-        content += HoverContent.Code(value, language)
-    }
-
-    fun code(language: String = "groovy", block: () -> String) {
-        content += HoverContent.Code(block(), language)
-    }
-
-    fun markdown(value: String) {
-        content += HoverContent.Markdown(value)
-    }
-
-    fun section(title: String, block: HoverBuilder.() -> Unit) {
-        val sectionContent = HoverBuilder().apply(block).build()
-        content += HoverContent.Section(title, sectionContent)
-    }
-
-    fun list(vararg items: String) {
-        content += HoverContent.List(items.toList())
-    }
-
-    fun list(items: kotlin.collections.List<String>) {
-        content += HoverContent.List(items)
-    }
-
-    fun keyValue(vararg pairs: Pair<String, String>) {
-        content += HoverContent.KeyValue(pairs.toList())
-    }
-
-    fun keyValue(pairs: kotlin.collections.List<Pair<String, String>>) {
-        content += HoverContent.KeyValue(pairs)
-    }
-
-    fun build(): kotlin.collections.List<HoverContent> = content.toList()
-
-    fun render(): String = content.joinToString("\n\n") { it.render() }
-}
-
-/**
- * DSL function for building hover content
- */
-fun hover(block: HoverBuilder.() -> Unit): Hover {
-    val content = HoverBuilder().apply(block).render()
-
-    val markupContent = MarkupContent().apply {
-        kind = MarkupKind.MARKDOWN
-        value = content
-    }
-
-    return Hover().apply {
-        contents = org.eclipse.lsp4j.jsonrpc.messages.Either.forRight(markupContent)
-    }
-}
 
 /**
  * Extension functions for formatting specific AST node types
